requirements.txt
# Requirements for TensorFlow Federated development.
click==7.1.2
urllib3==1.26.4
absl-py==0.11.0
boto3==1.17.67
cryptography==3.3.1
lmdb==1.0.0
grpclib==0.4.1
protobuf==3.14.0
tensorflow==2.4.0
cgroupspy==0.1.6
watchdog==0.10.3
torch==1.7.1
torchvision==0.8.2
nvidia-ml-py==11.450.51
diffiehellman==0.13.3
secret-sharing==0.2.7
kubernetes==12.0.1
psutil==5.8.0
tornado==6.1
pymongo==3.11.3
psycopg2==2.7.7
pycryptodome==3.10.1
pyOpenSSL==20.0.1
